requirements.txt
Pillow>=9.0.0
ExifRead>=3.0.0